requirements.txt

numpy
scipy
scikit-learn
imbalanced-learn
pyyaml
pandas
matplotlib
seaborn
lightgbm
mlflow
google-cloud-storage

